- Change all *Benchmark functions to return measured benchmark in double. If there are two processes in one benchmark, use receive one. - Each *Benchmark functions must return only bandwidth. Do not return exit code. Instead of non 0 returning from them, use CHECK macro inside functions. - Show the result in bandwidth.cc not to the log but to the stdout in GiByte/sec unit.

+++ fifo_benchmark.h
@@ -2,5 +2,5 @@
 
 #include <cstdint>
 
-int RunFifoBenchmark(int num_iterations, int num_warmups, uint64_t data_size,
-                     uint64_t buffer_size);
+double RunFifoBenchmark(int num_iterations, int num_warmups, uint64_t data_size,
+                        uint64_t buffer_size);
