+++ memcpy_benchmark.h
@@ -2,4 +2,5 @@
 
 #include <cstdint>
 
-int RunMemcpyBenchmark(int num_iterations, int num_warmups, uint64_t data_size);
+double RunMemcpyBenchmark(int num_iterations, int num_warmups,
+                          uint64_t data_size);
